requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
catboost
xgboost
lightgbm
streamlit
requests
python-dotenv
meteostat